+++ src/app/cart/cart.module.ts
@@ -1,14 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { CartListComponent } from './components/cart-list/cart-list.component';
-
-
+import { CartItemComponent } from './components/cart-item/cart-item.component';
 
 @NgModule({
-  declarations: [CartListComponent],
-  imports: [
-    CommonModule
-  ],
-  exports: [CartListComponent]
+  declarations: [CartListComponent, CartItemComponent],
+  imports: [CommonModule],
+  exports: [CartListComponent],
 })
-export class CartModule { }
+export class CartModule {}