+++ src/app/products/components/product/product.component.ts
@@ -1,18 +1,22 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+
+import { ProductModel } from '../../models/product.model';
+import { ProductsService } from '../../services/products.service';
 
 @Component({
   selector: 'app-product',
   templateUrl: './product.component.html',
-  styleUrls: ['./product.component.scss']
+  styleUrls: ['./product.component.scss'],
 })
 export class ProductComponent implements OnInit {
+  @Input() product!: ProductModel;
+  @Output() addToCart = new EventEmitter<ProductModel>();
 
-  constructor() { }
+  constructor(private productsService: ProductsService) {}
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
-  onAddToCart() {
-    console.log('Add to cart');
+  onAddToCart(): void {
+    this.addToCart.emit(this.product);
   }
 }