+++ src/app/products/services/products.service.ts
@@ -4,36 +4,36 @@ import { Category } from '../models/category.enum';
 import { ProductModel } from '../models/product.model';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class ProductsService {
   productsList: ProductModel[] = [
     {
       name: 'Apple iPhone 12 Pro Max',
-      description: 'Lorem ipsum dolor sit amet',
+      description: '2020 Wi-Fi 256GB Silver (MXAU2)',
       price: 50_000,
       category: Category.iPhone,
       isAvailable: true,
       colors: ['Graphite', 'Pacific Blue', 'Silver', 'Gold'],
-      memory: [128, 256, 512]
+      memory: [128, 256, 512],
     },
     {
       name: 'Apple iPhone 12',
-      description: 'Lorem ipsum dolor sit amet',
+      description: '2020 Wi-Fi 256GB Silver (MXAU2)',
       price: 38_000,
       category: Category.iPhone,
       isAvailable: true,
       colors: ['Graphite', 'Pacific Blue', 'Silver', 'Gold', 'Rose Gold'],
-      memory: [128, 256, 512]
+      memory: [128, 256, 512],
     },
     {
       name: 'Apple iPhone 8 Plus',
-      description: 'Lorem ipsum dolor sit amet',
+      description: '2020 Wi-Fi 256GB Silver (MXAU2)',
       price: 38_000,
       category: Category.iPhone,
       isAvailable: false,
       colors: ['Yellow', 'Red', 'Silver', 'Gold', 'Rose Gold'],
-      memory: [128, 256, 512]
+      memory: [128, 256, 512],
     },
     {
       name: 'Apple iPad Pro 4 12.9',
@@ -42,14 +42,13 @@ export class ProductsService {
       category: Category.iPad,
       isAvailable: true,
       colors: ['Graphite', 'Pacific Blue', 'Silver', 'Gold'],
-      memory: [128, 256, 512]
-    }
+      memory: [128, 256, 512],
+    },
   ];
 
-  constructor() { }
+  constructor() {}
 
   getProducts() {
-    console.log('Get products');
     return this.productsList;
   }
 }