+++ changelog.md
@@ -11,3 +11,21 @@
 - ProductsService
 - CartListComponent
 - CartService
+
+# Task 2. Introduction
+
+## October 7th 2021
+### Added:
+- Orders & Shared Modules
+- Bootstrap, Icons
+- Add to cart functionality
+- Total Price, Total Quantity
+- CartItemComponent
+- СartItemComponent OnPush strategy
+- Header Component: #appTitle
+- Highlight Directive
+
+### Changed:
+- ProductListComponent
+- ProductItemComponent
+- CartService